🚚 공통 위젯 분리
공통 위젯 분리
