+++ lib/ui/onboarding/views/onboarding_duration_view.dart
@@ -7,7 +7,9 @@ import '../../../theme/luckit_typos.dart';
 import '../../common/widgets/custom_text_field.dart';
 import '../onboarding_state.dart';
 import '../onboarding_view_model.dart';
+import '../widgets/error_text_widget.dart';
 import '../widgets/onboarding_bottom_button.dart';
+import '../widgets/onboarding_date_input_field_widget.dart';
 import '../widgets/onboarding_layout.dart';
 import '../widgets/onboarding_top_widget.dart';
 import '../widgets/suggested_duration_widget.dart';
@@ -210,40 +212,5 @@ class OnboardingDurationInputWidget extends StatelessWidget {
             ),
           ),
         ],
-      );
-}
-
-// TODO(dodo): errorText 개선
-class OnboardingDateInputField extends StatelessWidget {
-  final Function(String)? onChanged;
-  final String hintText;
-  final String? errorText;
-
-  const OnboardingDateInputField({
-    required this.hintText,
-    this.onChanged,
-    this.errorText,
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) => CustomTextField(
-        hintText: hintText,
-        errorText: errorText,
-        onChanged: onChanged,
-        fieldHeight: 40.0,
-        fillColor: LuckitColors.white,
-        enabledBorderColor: LuckitColors.gray20,
-        focusedBorderColor: LuckitColors.main,
-        textInputType: TextInputType.number,
-        textStyle: LuckitTypos.suitR20.copyWith(
-          color: LuckitColors.gray80,
-          height: 0.0,
-        ),
-        hintTextStyle: LuckitTypos.suitR20.copyWith(
-          color: LuckitColors.gray20,
-          height: 0.0,
-        ),
-        textAlign: TextAlign.center,
       );
 }
