+++ lib/ui/onboarding/widgets/suggested_duration_widget.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../../theme/luckit_colors.dart';
 import '../../../theme/luckit_typos.dart';
+import '../../common/widgets/rounded_grey_text_widget.dart';
 
 class SuggestedDurationWidget extends StatelessWidget {
   final bool isChecked;
@@ -14,20 +15,8 @@ class SuggestedDurationWidget extends StatelessWidget {
   });
 
   @override
-  Widget build(BuildContext context) => Container(
-      padding: const EdgeInsets.symmetric(
-        horizontal: 8.0,
-        vertical: 4.0,
-      ),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(12.0),
-        color: LuckitColors.gray10,
-      ),
-      child: Text(
-        '추천 기간 $duration',
-        style: LuckitTypos.suitR10.copyWith(
-          color: isChecked ? LuckitColors.gray60 : LuckitColors.gray40,
-        ),
-      ),
-    );
+  Widget build(BuildContext context) => RoundedGreyTextWidget(
+        label: '추천 기간 $duration',
+        activated: isChecked,
+      );
 }
